Updated Analysis
Analysis will now continue even if there is an error. The first default analysis script doesn't even exist, and thus errors out right away. This way it will continue to the next analysis script. (i haven't tested against 10's of thousands of devices, but the analysis bit is rather quick in the scheme of things, and to fail right away and just kill the whole script is a bit ridiculous)